不要なファイルの削除とタイマー機能の実装をトップページに移動
