+++ frontend/src/app/components/auth/LoginForm.tsx
@@ -81,12 +81,6 @@ export default function LoginForm() {
                 <div className="grid gap-2">
                   <div className="flex items-center">
                     <Label htmlFor="password">パスワード</Label>
-                    <a
-                      href="#"
-                      className="ml-auto inline-block text-sm underline-offset-4 hover:underline"
-                    >
-                      パスワードをお忘れの方
-                    </a>
                   </div>
                   <Input
                     id="password"
